Moved the DummyNetworkHandler to its own package, fixed problems with DeviceTest
