+++ Hestia/app/src/main/java/hestia/backend/models/Device.java
@@ -30,13 +30,6 @@ public class Device {
     private ArrayList<Activator> activators;
     private NetworkHandler handler;
 
-    public Device(String deviceId, String name, String type, ArrayList<Activator> activator) {
-        this.deviceId = deviceId;
-        this.name = name;
-        this.type = type;
-        this.activators = activator;
-    }
-
     public Device(String deviceId, String name, String type, ArrayList<Activator> activator, NetworkHandler handler) {
         this.deviceId = deviceId;
         this.name = name;
